test(mcp): fix three real defects the integration run exposed
- tool_call_reveal seeded conversations.active_branch_id BEFORE the branch row
  existed (FK violation); adopt the branch after creating it, as every other
  integration test and showcase.sql already do
- tool_call_lookup bound random UUIDs to mcp_tool_calls.message_id, which is an
  FK to messages; seed a real message per id
- stream_tool_timing asserted the model makes EXACTLY one tool call. That is an
  assertion about model behaviour, not about the wire contract under test (it
  made 9). Assert >= 1; every frame is still checked, so more calls is a
  stronger assertion, not a weaker one
- jsonb_array_length returns INT4, decoded as i64

+++ src-app/server/tests/chat/showcase_seed_rail_test.rs
@@ -336,8 +336,9 @@ async fn showcase_seed_is_idempotent_and_every_tool_use_is_paired() {
     }
 
     // The artifact-producing run really carries resource_links.
+    // `jsonb_array_length` returns INT4; cast so the decode matches the binding.
     let artifact_links: i64 = sqlx::query_scalar(
-        "SELECT jsonb_array_length(content->'resource_links') FROM message_contents \
+        "SELECT jsonb_array_length(content->'resource_links')::bigint FROM message_contents \
          WHERE content_type = 'tool_result' AND content->>'tool_use_id' = 'toolu_rail_artifacts'",
     )
     .fetch_one(&pool)
